Extract our Mechanism class
