+++ src/Mechanisms/HandleRequests/HandleRequests.php
@@ -5,16 +5,12 @@ namespace Livewire\Mechanisms\HandleRequests;
 use Illuminate\Support\Facades\Route;
 use Livewire\Features\SupportScriptsAndAssets\SupportScriptsAndAssets;
 
+use Livewire\Mechanisms\Mechanism;
 use function Livewire\trigger;
 
-class HandleRequests
+class HandleRequests extends Mechanism
 {
     protected $updateRoute;
-
-    function register()
-    {
-        app()->singleton($this::class);
-    }
 
     function boot()
     {
