+++ src/Mechanisms/Mechanism.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Livewire\Mechanisms;
+
+use Illuminate\Support\Facades\Blade;
+use Illuminate\Support\Str;
+use Livewire\LivewireServiceProvider;
+
+abstract class Mechanism
+{
+    protected $singleton = true;
+
+    function register($provider)
+    {
+        if ($this->singleton) app()->singleton(static::class, $this);
+    }
+
+    function boot()
+    {
+        //
+    }
+}
