+++ src/Mechanisms/PersistentMiddleware/PersistentMiddleware.php
@@ -2,12 +2,13 @@
 
 namespace Livewire\Mechanisms\PersistentMiddleware;
 
+use Livewire\Mechanisms\Mechanism;
 use function Livewire\on;
 use Illuminate\Support\Str;
 use Livewire\Drawer\Utils;
 use Livewire\Mechanisms\HandleRequests\HandleRequests;
 
-class PersistentMiddleware
+class PersistentMiddleware extends Mechanism
 {
     protected static $persistentMiddleware = [
         \Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful::class,
@@ -22,11 +23,6 @@ class PersistentMiddleware
 
     protected $path;
     protected $method;
-
-    function register()
-    {
-        app()->singleton($this::class, fn () => $this);
-    }
 
     function boot()
     {
